requirements.txt
av<=10.0.0
cvzone<=1.5.6
face_recognition<=1.3.0
matplotlib<=3.7.2
mediapipe<=0.10.2
numpy<=1.24.0
opencv_contrib_python<=4.8.0.74
opencv_python<=4.8.0.74
pandas<=2.0.3
Pillow<=10.0.0
scipy<=1.11.1
st_pages<=0.4.3
streamlit<=1.24.1
streamlit_webrtc<=0.45.1
tensorflow<=2.13.0
tensorflow_intel<=2.13.0
tqdm<=4.65.0
